+++ R/AllGenerics.R
@@ -20,7 +20,6 @@ setGeneric("removeReporters", function(object, ...) standardGeneric("removeRepor
 setGeneric("fromFile", function(object) standardGeneric("fromFile"))
 setGeneric("fromFile<-", function(object, value) standardGeneric("fromFile<-"))  ## This one should remain "private"
 
-setGeneric("quantify", function(object, ...) standardGeneric("quantify"))
 setGeneric("curveStats", function(object,reporters, ...) standardGeneric("curveStats"))
 setGeneric("purityCorrect", function(object,impurities, ...) standardGeneric("purityCorrect"))